feat: drop requirement for AD flag from the resolver
We do all the validation mandated by RFC 7344 ourselves
using dnspython library. Requiring AD is unnecessary and harmful
if you use CDS to upgrade to a newer algorithm, that is supported
dnspython but not by the resolver.

N. B.: the check for AD flag was wrong and never worked. Comparing
the result of binary operation to a boolean value using `is` is a bad
idea.