Version 0.9: implement DNSSEC policies
This release fixes compatibility issues with dnspython 2.3.0 that
introduces DNSSEC policies.

We do follow RFC 8624 policies for continuity check and signed-by-KSK
check. This means we refuse to roll to a DS record pointing to a
deprecated algorithm and/or hashing function. We still allow rolling
from them to some other supported algorithms.